Final Project Bookstore Application
The files included are all the necessary files to recreate the bookstore app. Download ionic through CLI and create an ionic tab project. Generate a few angular services that are able to interact with the tab's pages. This application is trying to represent a bookstore app. On the first tab, you are able to add books into a cart and have a few options with your desired cart. You are able to share the item, edit the item, and to delete the item in your cart. The second tab is the library tab. Show cases what is popular in the library. There is a FAB that once clicked gives the user to change the type of books to be displayed. There is also a search bar that helps users look for a book. The final tab is the profile tab. The intention is for the user to create an account to keep preferences and to share through social media and to leave reviews on books. There is another FAB that is a camera that gives users the ability to take a photo to post on their profile.